Changed all of GMRoute data members over to pointers.
Updated all GMRoute getters, setters, and engines to accomodate this.

Members are pointers so that they can mimic the NULL database value. This will also cause an explosion if someone tries to dereference a member when it is null. Don't do calculations with these bad boys unless you check them first!